+++ opencascade-hs/cpp/hs_BRepAdaptor_Curve.cpp
@@ -25,3 +25,11 @@ Handle(Geom_BSplineCurve) * hs_BRepAdaptor_Curve_bspline(BRepAdaptor_Curve * cur
 GeomAdaptor_Curve * hs_BRepAdaptor_Curve_curve(BRepAdaptor_Curve * curve){
     return new GeomAdaptor_Curve(curve->Curve());
 }
+
+double hs_BRepAdaptor_Curve_firstParameter(BRepAdaptor_Curve *curve){
+    return curve->FirstParameter();
+}
+
+double hs_BRepAdaptor_Curve_lastParameter(BRepAdaptor_Curve *curve){
+    return curve->LastParameter();
+}