+++ opencascade-hs/src/OpenCascade/BRepAdaptor/Curve.hs
@@ -6,6 +6,8 @@ module OpenCascade.BRepAdaptor.Curve
 , bezier
 , bspline
 , curve
+, firstParameter
+, lastParameter
 ) where
  
 import OpenCascade.BRepAdaptor.Types (Curve)
@@ -15,9 +17,10 @@ import qualified OpenCascade.Geom.Types as Geom
 import OpenCascade.GeomAbs.CurveType (CurveType)
 import qualified OpenCascade.TopoDS as TopoDS
 import Foreign.Ptr (Ptr)
-import Foreign.C (CInt (..))
+import Foreign.C (CInt (..), CDouble (..))
 import Data.Acquire (Acquire, mkAcquire)
 import OpenCascade.Handle (Handle)
+import Data.Coerce (coerce)
 import qualified OpenCascade.GeomAdaptor.Types as GeomAdaptor
 import qualified OpenCascade.GeomAdaptor.Internal.Destructors as GeomAdaptor.Destructors
 
@@ -47,3 +50,13 @@ foreign import capi unsafe "hs_BRepAdaptor_Curve.h hs_BRepAdaptor_Curve_curve" r
 
 curve :: Ptr Curve -> Acquire (Ptr GeomAdaptor.Curve)
 curve theCurve = mkAcquire (rawCurve theCurve) GeomAdaptor.Destructors.deleteCurve
+
+foreign import capi unsafe "hs_BRepAdaptor_Curve.h hs_BRepAdaptor_Curve_firstParameter" rawFirstParameter :: Ptr Curve -> IO CDouble
+
+firstParameter :: Ptr Curve -> IO Double
+firstParameter = coerce rawFirstParameter
+
+foreign import capi unsafe "hs_BRepAdaptor_Curve.h hs_BRepAdaptor_Curve_lastParameter" rawLastParameter :: Ptr Curve -> IO CDouble
+
+lastParameter :: Ptr Curve -> IO Double
+lastParameter = coerce rawLastParameter