+++ opencascade-hs/src/OpenCascade/ShapeConstruct/Internal/Destructors.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE CApiFFI #-}
+module OpenCascade.ShapeConstruct.Internal.Destructors
+( deleteCurve
+) where
+
+import OpenCascade.ShapeConstruct.Types
+import Foreign.Ptr 
+
+foreign import capi unsafe "hs_ShapeConstruct_Curve.h hs_delete_ShapeConstruct_Curve" deleteCurve :: Ptr Curve -> IO ()